+++ asdf/tags/core/complex.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from asdf import util
-from asdf.types import AsdfType
+from asdf._types import AsdfType
 
 
 class ComplexType(AsdfType):